+++ src/app/app-routing.module.ts
@@ -1,8 +1,14 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { ProductComponent } from './feature/product/components/product/product.component';
+import { Routes, RouterModule, ExtraOptions } from '@angular/router';
 import { AboutComponent, PathNotFoundComponent } from './layout';
+import { CartComponent } from './feature/cart/cart.component';
+import { CustomPreloadingStrategyService } from './feature/core/custom-preloading-strategy.service';
+import { FeedbacksComponent } from './feature/product/components';
 
+const extraOptions: ExtraOptions = {
+  preloadingStrategy: CustomPreloadingStrategyService,
+  enableTracing: true // Makes the router log all its internal events to the console.
+};
 
 const routes: Routes = [
   {
@@ -10,10 +16,17 @@ const routes: Routes = [
     component: AboutComponent
   },
   {
-    path: '',
+    path: 'product',
     redirectTo: '/home',
     pathMatch: 'full'
   },
+  {
+    path: 'cart',
+    loadChildren: () => import('./feature/cart/cart.module').then(m => m.CartModule),
+    data: {
+      preload: true
+    }
+  }, 
   {
     // The router will match this route if the URL requested
     // doesn't match any paths for routes defined in our configuration