+++ src/app/feature/cart/cart.component.ts
@@ -1,0 +1,21 @@
+import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+
+@Component({
+  selector: 'app-cart',
+  templateUrl: './cart.component.html',
+  styleUrls: ['./cart.component.css']
+})
+export class CartComponent implements OnInit {
+  constructor(
+    private router: Router
+  ) { }
+
+  ngOnInit() {
+  }
+
+  onCreateOrder() {
+    const link = ['/orders/add'];
+    this.router.navigate(link);
+  }
+}