+++ src/app/feature/cart/cart.module.ts
@@ -1,19 +1,22 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
-import { CartService } from '../shared/services/cart.service';
 import { CartListComponent } from './components/cart-list/cart-list.component';
-import { CartModel } from './models/cart.model';
 
 import { FormsModule } from '@angular/forms';
 import { SharedModule } from 'src/app/shared/shared.module';
+import { CartService } from './services/cart.service';
+import { CartComponent } from './cart.component';
+import { CartRoutingModule } from './cart-routing.module';
+import { OrderFormComponent } from './components/order-form/order-form.component';
 @NgModule({
-  imports: [CommonModule , FormsModule, SharedModule],
+  imports: [CommonModule, FormsModule, SharedModule, CartRoutingModule],
 
-  declarations: [
+  declarations: [CartComponent,
     CartItemComponent,
-    CartListComponent],
+    CartListComponent,
+    OrderFormComponent],
   providers: [CartService],
-  exports: [CartListComponent]
+  exports: [CartComponent]
 })
 export class CartModule { }