+++ src/app/feature/cart/models/order.model.ts
@@ -1,0 +1,8 @@
+import { CartModel } from './cart.model';
+
+export class OrderModel {
+  name: string;
+  phone: String;
+  constructor(
+    public products: CartModel[]) { }
+}