+++ src/app/feature/product/components/product-list/product-list.component.ts
@@ -1,9 +1,10 @@
-import { Component, OnInit, Output, EventEmitter} from '@angular/core';
+import { Component, OnInit, Output, EventEmitter } from '@angular/core';
 
 import { ProductModel } from '../../models/product.model';
-import { ProductService } from '../../../shared/services/products.service';
+import { ProductService } from '../../service/products.service';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
+import { CartService } from 'src/app/feature/cart/services/cart.service';
 
 @Component({
   selector: 'app-product-list',
@@ -12,25 +13,25 @@ import { Observable } from 'rxjs';
   providers: [ProductService]
 })
 export class ProductListComponent implements OnInit {
-  @Output() buyProd: EventEmitter<any> = new EventEmitter();
+ // @Output() buyProd: EventEmitter<any> = new EventEmitter();
 
   products: Observable<Array<ProductModel>>;
   constructor(private productsService: ProductService,
-    private router: Router) { }
+    private router: Router, private cartService: CartService) { }
 
   ngOnInit() {
     this.products = this.productsService.getProducts();
   }
 
   onBuy(prod: ProductModel): void {
-    this.buyProd.emit(prod);
+    this.cartService.buy(prod);
   }
-  onDetails(prod: ProductModel): void {
-    const link = ['/details', prod];
+  onDetails(prodId: number): void {
+    const link = ['/details',prodId];
     this.router.navigate(link);
   }
-  
-  onDisplayFeedbacks(prodId:number): void {
-    this.router.navigate([{ outlets: { feedbacks: ['feedbacks',prodId] } }]);
+
+  onDisplayFeedbacks(prodId: number): void {
+    this.router.navigate([{ outlets: { feedbacks: ['feedbacks', prodId] } }]);
   }
 }